+++ src/main/kotlin/io/github/realyusufismail/armourandtoolsmod/core/init/RecipeTypeInit.kt
@@ -18,9 +18,9 @@
  */ 
 package io.github.realyusufismail.armourandtoolsmod.core.init
 
-import io.github.realyusufismail.armourandtoolsmod.MOD_ID
-import io.github.realyusufismail.armourandtoolsmod.core.blocks.armour.CustomArmourCraftingTableRecipe
-import io.github.realyusufismail.armourandtoolsmod.core.blocks.tool.CustomToolCraftingTableRecipe
+import io.github.realyusufismail.armourandtoolsmod.ArmourAndToolsMod.ArmorAndToolsMod.MOD_ID
+import io.github.realyusufismail.armourandtoolsmod.blocks.armour.CustomArmourCraftingTableRecipe
+import io.github.realyusufismail.armourandtoolsmod.blocks.tool.CustomToolCraftingTableRecipe
 import net.minecraft.resources.ResourceLocation
 import net.minecraft.world.item.crafting.RecipeType
 import net.minecraftforge.registries.DeferredRegister